Issue #120 Add double click support (#121)
* Issue #120 Double click support

This is a requirement in order to implement inventory usage
for items like spynotes, scrolls, restoratives, etc.

I've added a class Gui::DoubleClickStateMachine. However
sending a native double click event incurs too much input latency.

Instead when an item is selected and is clicked again, we trigger 
its "use" scenario. This has been done for spynotes to begin with.