Closes #153 Display world objects that are not present in WLD files (#176)
* Closes #153 Display world objects that are not present in WLD files

Some objects are not present in the WLD files, these need to be manually
added to the map when this is detected. Additionally some only appear in
certain chapters. e.g. Rift gate in ch7. The model information is part of
the ContainerWorldLocation header so we have all the info we need to
place these on the map.

* Fix this for chapter 4 Gorath's chest