Closes #180 Fix buggy v1.02 dialog actions
v1.02 of BAK purported to fix a number of issues with v1.01,
unfortunately the patching process appears to have broken a number of
dialog actions, including many audio actions, as well as a few game
critical actions like the Fadamor's well, Prank's stone, and many
others.

I've added a list of dialog action replacements that come from v1.01 and
will be applied to the dialog actions of snippets as they are built.