Issue #207 Allow substituting higher res graphics (#215)
* Issue #207 Allow substituting higher res graphics

Added functionality to texture factory to search for substitute textures. 
Will look in the graphics mod dir for file names the same as the BMX but 
with a PNG extension. For multi-image BMX files it will look in the sub-folder
`INVSHP0.BMX/0.PNG` for the sub images. 

This affects textures used in the 2D and 3D view.

Also changed the image extractor to extract direct to PNG so that modders
can generate a directory structure and simply replace files as they go.